requirements.txt
requests~=2.28.2
selenium~=4.8.2
awscli
Pillow~=9.4.0
Flask~=2.2.3
Faker~=18.2.0